+++ noodles-vcf/src/header/record/value/map/contig/tag.rs
@@ -4,12 +4,14 @@ use crate::header::record::value::map::tag::{self, ID, IDX};
 
 const LENGTH: &str = "length";
 const MD5: &str = "md5";
+const URL: &str = "URL";
 
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
 pub enum Standard {
     Id,
     Length,
     Md5,
+    Url,
     Idx,
 }
 
@@ -21,6 +23,7 @@ impl AsRef<str> for Standard {
             Self::Id => ID,
             Self::Length => LENGTH,
             Self::Md5 => MD5,
+            Self::Url => URL,
             Self::Idx => IDX,
         }
     }
@@ -34,6 +37,7 @@ impl FromStr for Standard {
             ID => Ok(Self::Id),
             LENGTH => Ok(Self::Length),
             MD5 => Ok(Self::Md5),
+            URL => Ok(Self::Url),
             IDX => Ok(Self::Idx),
             _ => Err(()),
         }
@@ -49,6 +53,7 @@ mod tests {
         assert_eq!(Standard::Id.as_ref(), "ID");
         assert_eq!(Standard::Length.as_ref(), "length");
         assert_eq!(Standard::Md5.as_ref(), "md5");
+        assert_eq!(Standard::Url.as_ref(), "URL");
         assert_eq!(Standard::Idx.as_ref(), "IDX");
     }
 }